python-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.4
python-multipart==0.0.6
# LangChain dependencies - compatible versions
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0